requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aniso8601=8.0.0=py_0
attrs=20.2.0=pyh9f0ad1d_0
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38h32f6830_0
click=7.1.2=py_0
flask=1.1.2=py_0
flask-jsonpify=1.5.0=py_1000
flask-restplus=0.13.0=py38_0
flask-socketio=4.3.1=pyh9f0ad1d_0
freetype=2.10.2=hd328e21_0
icu=67.1=h33f27b4_0
importlib-metadata=2.0.0=py38h32f6830_0
importlib_metadata=2.0.0=0
intel-openmp=2020.2=254
itsdangerous=1.1.0=py_0
jinja2=2.11.2=py_0
jpeg=9d=he774522_0
jsonschema=3.2.0=py38h32f6830_1
libblas=3.8.0=8_h8933c1f_netlib
libcblas=3.8.0=8_h8933c1f_netlib
libclang=10.0.1=default_hf44288c_1
liblapack=3.8.0=8_h8933c1f_netlib
liblapacke=3.8.0=8_h8933c1f_netlib
libopencv=4.4.0=py38_2
libpng=1.6.37=ha81a0f5_2
libtiff=4.1.0=h885aae3_6
libwebp-base=1.1.0=hfa6e2cd_3
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
opencv=4.4.0=py38_2
openssl=1.1.1h=he774522_0
pip=20.2.3=py38_0
py-opencv=4.4.0=py38h23f93f0_2
pyrsistent=0.17.3=py38h1e8a9f7_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
python-engineio=3.11.2=py_0
python-socketio=4.5.1=pyh9f0ad1d_0
python_abi=3.8=1_cp38
pytz=2020.1=pyh9f0ad1d_0
qt=5.12.6=hb2cf2c5_0
setuptools=49.6.0=py38_1
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_1
zipp=3.3.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h1f3a1b7_2
